lint: adopt ruff with everything on, and fix what it found
select = ["ALL"], then turn off only what is wrong for this codebase and
say why in the config rather than in a commit nobody will find again. The
reasoning here lives in block comments above the code it explains, so the
docstring rules would only move it somewhere worse; the solver uses the
notation of the literature, so renaming A and b to satisfy N803 would
make it harder to check against the paper.

Where a default threshold was simply arbitrary for this domain it is
tuned, not suppressed. Building a fault dictionary genuinely needs six
independent inputs, and bundling them into a config object to get under
max-args would hide the signature rather than simplify it.

The real finds, all fixed rather than ignored:

  - align was a positional boolean on seven public functions. Now
    keyword-only, which is what every caller already did. factors stays
    positional, since callers pass it that way and breaking them to
    please a linter is not an improvement.
  - os.path throughout, now pathlib, including one open() that was never
    closed.
  - 67 printf-style format strings, now f-strings.
  - bench.py's main() was 86 statements doing four unrelated jobs. Split
    into parse_args, bench_devices and report, which also gives the
    verdict prose a name.

Every conversion was checked by diffing the four suites' output against
the run before any of this started. It is byte-identical.

Formatting is deliberately not enforced. The wrapping here is hand-set to
keep the prose readable, and ruff format would reflow all of it.

Also wires ruff into CI beside the tests.

+++ netfault/measure.py
@@ -22,7 +22,7 @@ def tone(f, seconds, fs, amplitude=0.5):
     neighbours, which is what lets level_at() be a single bin rather than
     a windowed estimate.
     """
-    n = max(int(round(seconds * fs / (fs / f))) , 1) * int(round(fs / f))
+    n = max(round(seconds * fs / (fs / f)) , 1) * round(fs / f)
     t = np.arange(n) / float(fs)
     return amplitude * np.sin(2.0 * np.pi * f * t)
